only download if not exists

+++ vagrant/scripts/install-jb-toolbox.sh
@@ -1,19 +1,23 @@
-#!/bin/bash
+#!/usr/bin/env bash
 
 JBTBURL=https://download-cdn.jetbrains.com/toolbox/jetbrains-toolbox-1.21.9712.tar.gz
+TOOLLOC=/bin/jetbrains-toolbox
 
-pushd /tmp/
+if [[ ! -f $TOOLLOC ]]; then
+    pushd /tmp/
 
-curl $JBTBURL > toolbox.tar.gz
+        curl $JBTBURL > toolbox.tar.gz
 
-gunzip toolbox.tar.gz
-tar -xf toolbox.tar
+        gunzip toolbox.tar.gz
+        tar -xf toolbox.tar
 
-pushd jetbrains-toolbox-*/
-chmod +x jetbrains-toolbox
-mv jetbrains-toolbox /bin/
+        pushd jetbrains-toolbox-*/
+            chmod +x jetbrains-toolbox
+            mv jetbrains-toolbox /bin/
+        popd
+    popd
 
-popd
-popd
-
-jetbrains-toolbox &
+    jetbrains-toolbox &
+else
+    echo "Already exists: $TOOLLOC"
+fi
